% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{add_bib}
\alias{add_bib}
\title{Add .bib file}
\usage{
add_bib(database, dataset)
}
\arguments{
\item{database}{Name of the database dataset is a part of}

\item{dataset}{Name of the dataset}
}
\value{
A .bib template saved to the data-raw folder to be completed
}
\description{
Add .bib file template to help cite the datasets in a qPackage
}
\examples{
\dontrun{
add_bib("states", "COW")
}
}
