% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_action}
\alias{code_action}
\title{Code Actions for Titles}
\usage{
code_action(title)
}
\arguments{
\item{title}{A character vector of treaty title}
}
\value{
A character vector with 2 letter action abbreviations
for bilateral treaties
}
\description{
Identifies actions performed by agreements
so that bileteral treaties can be better distinguished.
}
\details{
Actions of agreements help differentiate date duplicates
in the same dataset as different treaties.
Actions refers to verbs coded from title.
For the complete list of actions and their 2 letter abbreviations
please refer to the actions list by running the function without
argument (i.e. \code{code_action()}).
}
