% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{code_actions}
\alias{code_actions}
\title{Code Actions from agreement titles}
\usage{
code_actions(title)
}
\value{
The action taken from agreement title
}
\description{
Code Actions from agreement titles
}
\examples{
code_actions(qEnviron::agreements$IEADB$Title)
}
