% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_agreements}
\alias{code_agreements}
\title{Code Agreement Titles}
\usage{
code_agreements(dataset = NULL, title, date)
}
\arguments{
\item{dataset}{dataset name.
If provided without a title and date column,
the function finds title and date conforming columns
in the dataset automatically if available.}

\item{title}{title column variable.
Ideally, title variable should come from a qPackage database/dataset
for which strings were standardised with \code{standardise_titles()}}

\item{date}{date column variable
Ideally, date variable should come from a qPackage database/dataset
for which dates were standardised with \code{standardise_dates()}}
}
\value{
a character vector with the qIDs
}
\description{
Creates an ID that contains information on the
parties to an agreement, the type of agreement, the date
and the linkage to other agreements in the dataset.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
code_agreements(dataset = IEADB)
code_agreements(title = IEADB$Title, date = IEADB$Signature)
}
