% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{code_area}
\alias{code_area}
\title{Code Agreement Area}
\usage{
code_area(title)
}
\arguments{
\item{title}{}
}
\value{
The region of the agreement
}
\description{
Code Agreement Area
}
\examples{
code_area(qEnviron::agreements$IEADB$Title)
}
