% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_domain}
\alias{code_domain}
\title{Code Domain from agreement titles}
\usage{
code_domain(title)
}
\arguments{
\item{title}{Treaty titles}
}
\value{
The domain taken from agreement title
}
\description{
Code Domain from agreement titles
}
\examples{
\donttest{
title <- sample(manyenviron::agreements$IEADB$Title, 30)
code_domain(title)
}
}
