% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_entity}
\alias{code_entity}
\title{Code Agreement Entity}
\usage{
code_entity(title, destination = NULL)
}
\arguments{
\item{title}{Treaty titles}

\item{destination}{Do you want regions or continents
returned instead of location names?
By default NULL.
Options include continent and region.}
}
\value{
The region of the agreement
}
\description{
Code Agreement Entity
}
\examples{
\donttest{
title <- head(qEnviron::agreements$IEADB$Title)
code_entity(title)
}
}
