% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_grounds.R
\name{code_grounds}
\alias{code_grounds}
\title{Code Treaty Termination Type}
\usage{
code_grounds(title, text = NULL)
}
\arguments{
\item{title}{title column variable}

\item{text}{treaty text column variable}
}
\value{
a character vector with the termination clause type
of the treaty
}
\description{
Creates a column with the type of termination
of the treaty.
}
\examples{
GNEVAR <- dplyr::slice_sample(qEnviron::agreements$GNEVAR, n = 200)
GNEVAR$Term_Type <- code_grounds(GNEVAR$Title)
}
