% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_issue}
\alias{code_issue}
\title{Code agreement issues for date duplicates}
\usage{
code_issue(title, date)
}
\arguments{
\item{title}{A character vector of treaty title}

\item{date}{A date variable}
}
\value{
A character vector with 2 letter issue abbreviations for date duplicates
}
\description{
Identifies issues to which agreements belong to for agreements
signed in the same day.
}
\details{
Issues are smaller levels than topic or domain.
this is important for differentiating date duplicates as different
treaties signed in the same day often concern the same broader topic.
If the water is a topic oceans, rivers and lakes are issues.
For the complete list of issues and their 2 letter abbreviations
please refer to the issues list available in sysdata.
}
