% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_known_agreements}
\alias{code_known_agreements}
\title{Code Known Agreements Abbreviation}
\usage{
code_known_agreements(title)
}
\arguments{
\item{title}{A character vector of treaty title}
}
\value{
A character vector of abbreviation of known treaties
}
\description{
Some agreements have known abbreviations that facilitate their identification.
}
\details{
The function identifies agreements that match the list of known
agreements with their titles, abbreviations and signature dates and
substitutes the known titles for abbreviations.For the complete list of
known agreements coded for and their respective abbreviations please run
the function without an argument (i.e. \code{code_known_agreements()}).
}
\examples{
\dontrun{
IEADB <- dplyr::slice_sample(manyenviron::agreements$IEADB, n = 10)
code_known_agreements(IEADB$Title)
}
}
