% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{code_lineage}
\alias{code_lineage}
\title{Code lineage from agreement titles}
\usage{
code_lineage(database)
}
\arguments{
\item{database}{A database from the many packages ecosystem.}
}
\value{
A list of lineages that combines agreement area
and agreement action.
}
\description{
Code lineage from agreement titles
}
\examples{
code_lineage(qEnviron::agreements)
}
