% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termination.R
\name{code_term}
\alias{code_term}
\title{Code Treaty Termination Type}
\usage{
code_term(title, text = NULL)
}
\arguments{
\item{title}{title column variable}

\item{text}{treaty text column variable}
}
\value{
a character vector with the termination clause type
of the treaty
}
\description{
Creates a column with the type of termination
of the treaty.
}
\examples{
\donttest{
sample <- dplyr::slice_sample(manyenviron::texts$AGR_TXT, n = 200)
code_termination(sample$Title)
code_termination(sample$Title, sample$Text)
}
}
