% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_term.R
\name{code_term_date}
\alias{code_term_date}
\title{Code treaty termination date}
\usage{
code_term_date(title, text = NULL)
}
\arguments{
\item{title}{title column variable.}

\item{text}{treaty text column variable}
}
\value{
a character vector with the term date of the
treaty
}
\description{
Creates a column with the date of termination
of the treaty.
}
\examples{
\dontrun{
GNEVAR$Term_Date <- code_term_date(GNEVAR$Title)
}
}
