% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_term.R
\name{code_term_type}
\alias{code_term_type}
\title{Code treaty termination type}
\usage{
code_term_type(title, text = NULL)
}
\arguments{
\item{title}{title column variable}

\item{text}{treaty text column variable}
}
\value{
a character vector with the termination clause type
of the treaty
}
\description{
Creates a column with the type of termination
of the treaty.
}
\examples{
\dontrun{
GNEVAR$Term_Type <- code_term_type(GNEVAR$Title)
}
}
