% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{condense_qID}
\alias{condense_qID}
\title{Condense similar qIDs}
\usage{
condense_qID(database = NULL, var = NULL)
}
\arguments{
\item{database}{A qPackage database}

\item{var}{Two or more qID variables}
}
\value{
A dataframe of qID and qID references
}
\description{
Different qIDs generated for different datasets
might have minor differences in terms of acronym or linkage.
Some minor differences in qIDs could mean different qIDS
in different datasets actually refer to the same agreement.
The function finds these occurences and returns the
first qID argument entered as a replacement.
}
\examples{
data1 <- data.frame(qID = c("CPV-PRT[FSD]_1980A",
"CPV-PRT[FSD]_1990P:FSD_1980A",
"TD06LJ_1981A", "RAMSA_1971A", "WIIEWH_1982P"))
data2 <- data.frame(qID = c("TD06LJ_1981A", "RAMSA_1971A",
"WIIEWH_1982P:RAMSA_1971A",
"PRTRPC_1976A", "PRTRPC_1983E1:PRTRPC_1976A"))
qID_ref <- condense_qID(var = c(data1$qID, data2$qID))
data1 <- merge(data1, qID_ref)
data2 <- merge(data2, qID_ref)
}
