% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pdf.R
\name{convert_pdf}
\alias{convert_pdf}
\title{Get and convert PDFs to text}
\source{
https://stackoverflow.com/questions/71064939/how-to-check-if-pdf-is-scanned-image-or-contains-text-in-r
}
\usage{
convert_pdf(path)
}
\arguments{
\item{path}{The file(s) path.
The path can also be specified to a folder with multiple PDF files,
in that case all PDF files in the folder will be converted.}
}
\value{
A list with the converted texts.
}
\description{
Get and convert PDFs to text
}
\details{
If PDF file is not in "readable" format, function attempts to apply
optical character recognition (OCR) to convert these to text.
}
\examples{
\dontrun{
convert_pdf(path = getwd())
}
}
