% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_titles.R
\name{correct_words}
\alias{correct_words}
\title{Standardises some words spelling}
\usage{
correct_words(s)
}
\arguments{
\item{s}{A list of character vector}
}
\value{
A list of character vector with the words changed
}
\description{
Change some words spelling, specifically those that can vary
from one text to another.
}
\examples{
\dontrun{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 100)
IEADB$Title <- correct_words(IEADB$Title)
}
}
