% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_setup.R
\name{create_directory}
\alias{create_directory}
\title{Helper function from usethis:::create_directory()}
\usage{
create_directory(path)
}
\arguments{
\item{path}{path to directory}
}
\value{
a path to the directory
}
\description{
Helper function from usethis:::create_directory()
}
