% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manypkgs-defunct.R
\name{defunct}
\alias{defunct}
\alias{standardise_dates}
\alias{standardize_dates}
\title{Functions that have been renamed, superseded, or are no longer working}
\usage{
standardise_dates(...)

standardize_dates(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\section{Functions}{
\itemize{
\item \code{standardise_dates()}: Deprecated on 2022-05-17.
This wrapper function is deprecated and its functionality is included in the
original \code{make_messydates()} function. Please refer to its documentation
for more details about the new implementation.

\item \code{standardize_dates()}: Deprecated on 2022-05-17.
This wrapper function is deprecated and its functionality is included in the
original \code{make_messydates()} function. Please refer to its documentation
for more details about the new implementation.

}}
\keyword{internal}
