% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{export_data}
\alias{export_data}
\title{Adding datasets to the many packages universe}
\usage{
export_data(..., database, URL = NULL)
}
\arguments{
\item{...}{Unquoted name of the dataset object to save.}

\item{database}{Quoted name of any existing database or of the database to
be created.}

\item{URL}{Website URL to the source of a dataset.
Null by default.
Please make sure to fill in documentation file by hand if not provided.}
}
\value{
This function saves the dataset to the named database,
silently creates a set of tests for this dataset,
and creates and opens documentation for the dataset.
}
\description{
Save a cleaned data object, consistent with the manydata universe,
ready to be lazy-loaded and create scripts for documenting and
testing that object within the new package.
}
\details{
The function creates a data directory, if nonexistent, and
saves cleaned data. The functions also creates a script for testing
the cleaned data and make sure it complies with manydata requirements.
As well, it creates a documentation script to help documenting data
sources and describing variables. Note that users need to provide a \code{.bib}
file for citation purposes alongside their dataset in the corresponding
\code{data-raw} subfolder.
}
\examples{
\dontrun{
export_data(COW, database = "states",
URL = "https://correlatesofwar.org/data-sets/state-system-membership")
}
}
