% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{fuzzy_agreements_bilateral}
\alias{fuzzy_agreements_bilateral}
\title{Helper function for fuzzy matching bilateral agreements}
\usage{
fuzzy_agreements_bilateral(treatyID)
}
\arguments{
\item{treatyID}{treatyID variable created with \code{manypkgs::code_agreements()}}
}
\value{
A data frame with acronyms and treatyID matches without linkages
}
\description{
Fuzzy match agreements' acronyms for bilateral treaties
}
\details{
Extracts agreement acronyms from treatyIDs and fuzzy
match their similarities.
See \code{manypkgs::code_acronym()} for more details on acronyms.
}
