% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{fuzzy_agreements_multilateral}
\alias{fuzzy_agreements_multilateral}
\title{Helper function for fuzzy matching multilateral agreements}
\usage{
fuzzy_agreements_multilateral(qID)
}
\arguments{
\item{qID}{qID variable created with \code{manypkgs::code_agreements()}}
}
\value{
A data frame with acronyms and qID matches without linkages
}
\description{
Fuzzy match agreements' acronyms for multilateral treaties
}
\details{
Extracts agreement acronyms from qIDs and fuzzy
match their similarities.
See \code{manypkgs::code_acronym()} for more details on acronyms.
}
