% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{get_links}
\alias{get_links}
\title{Get links from treaty titles}
\usage{
get_links(database = NULL, dataset = NULL, treaty_type = "all")
}
\arguments{
\item{database}{A qPackage database}

\item{dataset}{A qPackage dataset}

\item{treaty_type}{The type of treaties to be returned.
By default, all.
Other options include bilateral or multilateral.}
}
\value{
A dataframe of agreements' qID and their linkages.
}
\description{
Treaties that modify, amend, or expand other treaties
usually specify so in the title.
We use the qIDs generated by \code{code_agreements()} and
\code{condense_qID()} to return a data frame of these for
for a dataset or a database.
}
\examples{
get_links(database = qEnviron::agreements)
get_links(dataset = qEnviron::agreements$IEADB)
}
