% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{get_memberships}
\alias{get_memberships}
\title{Get memberships' list}
\usage{
get_memberships(database, actor, id)
}
\arguments{
\item{actor}{A country variable}

\item{id}{A treaty ID variable}
}
\value{
A dataframe of treaty IDs and countries
part of the treaty
}
\description{
Memberships database have countries column and
treaty title column but information on the other
countries that are part to the treaties is often
missing.
The function allows to generate a dataframe of
treaty IDs and the countries part to the treaty.
}
\examples{
sample <- qEnviron::memberships$IEADB_MEM
get_memberships(actor = sample$CountryID, id = sample$qID_ref)
get_memberships(qEnviron::memberships)
}
