% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_export.R
\name{get_package_name}
\alias{get_package_name}
\title{Get the name of the package}
\usage{
get_package_name(path = getwd())
}
\arguments{
\item{path}{A string, if missing default is path to the working directory}
}
\value{
The name of the package
}
\description{
Get the name of the package
}
\examples{
\dontrun{
get_package_name()
}
}
