% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_treaty.R
\name{get_treaty}
\alias{get_treaty}
\title{Get treaty articles by number or match}
\usage{
get_treaty(
  t,
  article = NULL,
  match = NULL,
  memberships = NULL,
  termination = NULL
)
}
\arguments{
\item{t}{A list of treaty texts that has been "splitted" with
\code{manypkgs::split_treaty()}}

\item{article}{Either the "preamble" or an article number
across all treaties to be returned.
It can be left NULL if user intends to look for a word in
all treaties.}

\item{match}{A regex match for a word(s) or expression.
For multiple words, please use "|" to divide them.}

\item{memberships}{}

\item{termination}{}
}
\value{
A list of treaty sections of the same length
}
\description{
Once articles were split, you can access all the "preambles",
or first articles, or articles that contain certain words
with this function.
}
\examples{
\donttest{
t <- head(manyenviron::texts$AGR_TXT$Text)
t <- split_treaty(t)
get_treaty(t, article = "preamble")
get_treaty(t, article = 1)
get_treaty(t, match = "constitution")
get_treaty(t, article = "preamble", match = "unofficial")
get_treaty(t, memberships = TRUE)
get_treaty(t, termination = TRUE)
}
}
