% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_titles.R
\name{lingua}
\alias{lingua}
\title{Translate Strings}
\usage{
lingua(s, api_key, target_lang = "en", translate = TRUE)
}
\arguments{
\item{s}{A character string}

\item{api_key}{Google API key.
For more information please go to: https://cloud.google.com/translate/docs/setup}

\item{target_lang}{Which language would you like this translated to?
Please provide a two letter language abbreviation (e.g. "en" or "pt").
By default english.}

\item{translate}{Do you want strings to be translated?
By default TRUE.
If FALSE returns source language for each string.}
}
\value{
A character vector of the same length of original.
}
\description{
Tranlates strings in data and returns object of the same lenght as original string.
The function automacally identifies language from strings by rows
so that tranlations to target language are more accurate and allows
for text in multiple languages to be present in string.
}
