% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_metadata.R
\name{link_metadata}
\alias{link_metadata}
\title{Linking metadata of original dataset when adding datasets to the qPackage}
\usage{
link_metadata(dataset)
}
\arguments{
\item{dataset}{Name of the imported dataset.}
}
\value{
This function returns the original dataset with the additional
"metadata_orig" attribute.
}
\description{
This function saves original variable names and various other elements as
attributes to the imported objects to power the data_evolution() function
in qData. To be used in every data preparation script right after loading
the dataset in the environment.
}
\details{
This function returns the original dataset with the additional
"metadata_orig" attribute that will be used later by \code{data_evolution()}
to highlight changes that the dataset underwent during our import process.
}
\examples{
\dontrun{
link_metadata(COW)
}
}
