% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{pred_words}
\alias{pred_words}
\title{Predictable words list}
\usage{
pred_words()
}
\value{
A list of predicatble words removed
in code_linkages() and code_parties().
}
\description{
Predictable words list
}
