% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_membership_list}
\alias{retrieve_membership_list}
\title{Retrieve memberships' list}
\usage{
retrieve_membership_list(database, actor, id)
}
\arguments{
\item{database}{A database}

\item{actor}{An actor level variable (e.g. country)}

\item{id}{A treaty ID variable}
}
\value{
A dataframe of treaty IDs and actors
part of the treaty
}
\description{
Memberships database have actor column(s) and treaty column(s)
but information on the other countries that are party to the treaties
is often hard to read.
The function generates a dataframe of
treaty IDs and actors to a certain treaty.
}
\examples{
\dontrun{
sample <- manyenviron::memberships$IEADB_MEM
retrieve_membership_list(actor = sample$CountryID, id = sample$manyID)
retrieve_membership_list(manyenviron::memberships)
}
}
