% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{standardise_texts}
\alias{standardise_texts}
\title{Standardise treaty texts}
\usage{
standardise_texts(textvar)
}
\arguments{
\item{textvar}{A text variable.}
}
\value{
A list of treaty sections of the same length.
}
\description{
Standardise treaty texts by removing punctuation and markers,
while splitting these into articles and annexes.
}
\examples{
\dontrun{
standardise_texts(sample(manyenviron::texts$AGR_TXT$Text, 30))
}
}
