% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_treaty_text.R
\name{standardise_treaty_text}
\alias{standardise_treaty_text}
\alias{standardize_treaty_text}
\title{Standardise treaty texts}
\usage{
standardise_treaty_text(textvar)

standardize_treaty_text(textvar)
}
\arguments{
\item{textvar}{A text variable.}
}
\value{
A list of treaty sections of the same length.
}
\description{
The function standardises treaty texts by removing punctuation and markers,
while splitting these texts into articles and annexes.
}
\details{
Treaty texts are not always similar when imported to R.
Some treaty texts, for example, contain paragraph markers while others
come in one text chunk.
\code{standardise_treaty_text()} facilitates the cleaning and annotation of
these treaty texts so that information about clauses can be retrieved
at a later stage with the \code{retrieve_clauses()} function.
}
\examples{
\dontrun{
standardise_treaty_text(sample(manyenviron::agreements$HUGGO$MainText, 30))
}
}
