% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_titles.R
\name{standardise_words}
\alias{standardise_words}
\title{Standardises some words spelling}
\usage{
standardise_words(s)
}
\arguments{
\item{s}{A list of character vector}
}
\value{
A list of character vector with the words changed
}
\description{
Change some words spelling, specifically those that can vary
from one text to another.
}
\examples{
IEADB$Title <- standardise_words(IEADB$Title)
}
