% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_update.R
\name{update_package}
\alias{update_package}
\title{Updates packages of our data universe}
\usage{
update_package(package = NULL, name = NULL, path = getwd())
}
\arguments{
\item{package}{The name of the package to be updated, optional.
When not declared package name is extracted from description.}

\item{name}{A list of vectors giving the package author(s)' name(s).
Optional.
Authors(s)last name(s) and first name(s) are separated by a comma.}

\item{path}{The file path, optional.
If not specified, function gets the current working directory.}
}
\value{
Updated files in correct structure for packages.
}
\description{
As our packages evolve, expectations about package structures
as well as templates might also change.
This function updates files for packages setup with \code{setup_package()}.
}
\details{
Some of the files in the .github folder, such as workflow actions,
pull request templates, and issue templates, are automatically updated.
The function also asks developers if other files such as the package license,
contributing file, and code of conduct should be updated.
}
\examples{
\dontrun{
update_package("manystates")
}
}
