% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_accession.R
\name{code_accession_terms}
\alias{code_accession_terms}
\title{Get conditions and processes to accede memberships}
\usage{
code_accession_terms(t, title = NULL, accession = NULL)
}
\arguments{
\item{t}{A text variable}

\item{title}{A title variable.
If not declared, NULL by default.}

\item{accession}{The "condition" or "process"
for parties to accede a treaty.
If not declared, returns the "process".}
}
\value{
Either the conditions to be part of the treaty
or the different process steps to become a member.
}
\description{
The function allows to get information on the conditions
for parties to enter a treaty.
It could be limited to a entity area,
a certain domain,
or on State party nominations for new members.
The function also allows to get information on the
process by which a new member has to go through in
order to accede memberships
}
\examples{
\donttest{
m <- manyenviron::texts$AGR_TXT[100:300,]
code_accession_terms(m$Text, m$Title, accession = "condition")
code_accession_terms(m$Text, accession = "process")
code_accession_terms()
}
}
