% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_accession.R
\name{code_accession_terms}
\alias{code_accession_terms}
\title{Get conditions and processes to accede memberships}
\usage{
code_accession_terms(text, title = NULL, accession = NULL)
}
\arguments{
\item{text}{A text variable}

\item{title}{A title variable.
NULL by default.}

\item{accession}{The "condition" or "process" for parties to accede a treaty.
"process" by default.}
}
\value{
A list of the accession conditions or processes.
}
\description{
The function extracts information on the conditions for parties to become
part of a treaty or the different processes to become a treaty member.
}
\details{
The function helps provide metadata on accession to
treaties allowing researchers to connect and compare this information.
The processes relate to how new members accede treaty
and include treaty signature, ratification, notification, or unanimity vote.
The conditions might include an entity type or area, a certain issue-domain,
open membership for all, or nomination of new members.
}
\examples{
\donttest{
m <- manyenviron::texts$AGR_TXT[100:300,]
code_accession_terms(m$Text, m$Title, accession = "condition")
code_accession_terms(m$Text, accession = "process")
code_accession_terms()
}
}
