% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_agreements}
\alias{code_agreements}
\title{Code Agreement Titles}
\usage{
code_agreements(dataset = NULL, title, date)
}
\arguments{
\item{dataset}{dataset name.
If provided without a title and date variables,
the function finds title and date conforming
columns in the dataset.
The function "expects" that there are variables
named \code{Title} and \code{Signature} that they have been
standardised using \code{standardise_titles()} and
\code{standardise_dates()}, respectively.}

\item{title}{title variable.
The function "expects" that the variable has been
standardised using \code{standardise_titles()}}

\item{date}{date variable.
The function "expects" that the variable has been
standardised using \code{standardise_dates()}}
}
\value{
a character vector with the qIDs
}
\description{
Creates an ID that contains information on the
parties to an agreement, the type of agreement, the date
and the linkage to other agreements in the dataset.
}
\examples{
\donttest{
#' IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
code_agreements(dataset = IEADB)
code_agreements(title = IEADB$Title, date = IEADB$Signature)
}
}
