% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_agreements}
\alias{code_agreements}
\title{Code Agreement Titles}
\usage{
code_agreements(title, date, dataset = NULL)
}
\arguments{
\item{title}{title column variable.
Ideally, title variable should come from a qPackage database/dataset
for which strings were standardised with \code{standardise_titles()}}

\item{date}{date column variable
Ideally, date variable should come from a qPackage database/dataset
for which dates were standardised with \code{standardise_dates()}}

\item{dataset}{dataset name, optional.
If provided in place of title and date,
the function finds title and date conforming columns
in the dataset automatically if available.}
}
\value{
a character vector with the qIDs
}
\description{
Creates an ID that contains information on the
parties to an agreement, the type of agreement, the date
and the linkage to other agreements in the dataset.
}
\examples{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
IEADB$qID <- code_agreements(IEADB$Title, IEADB$Signature)
code_agreements(dataset = IEADB)
}
