% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{code_area}
\alias{code_area}
\title{Code Agreement Area}
\usage{
code_area(title, destination = NULL)
}
\arguments{
\item{title}{Treaty titles}

\item{destination}{Do you want regions or continents
returned instead of location names?
By default NULL.
Options include continent and region.}
}
\value{
The region of the agreement
}
\description{
Code Agreement Area
}
\examples{
title <- head(qEnviron::agreements$IEADB$Title)
code_area(title)
}
