% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_lineage}
\alias{code_lineage}
\title{Code lineage from agreement titles}
\usage{
code_lineage(title = NULL, database = NULL)
}
\arguments{
\item{title}{A title column for agreements}

\item{database}{A database from the many packages ecosystem.}
}
\value{
A list of lineages that combines agreement area
and agreement action.
}
\description{
Code lineage from agreement titles
}
\examples{
\donttest{
code_lineage(title = manyenviron::agreements$IEADB$Title)
code_lineage(database = manyenviron::agreements)
}
}
