% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lineage.R
\name{code_lineage}
\alias{code_lineage}
\title{Code lineage from agreement titles}
\usage{
code_lineage(title = NULL, datacube = NULL)
}
\arguments{
\item{title}{A title column for agreements}

\item{datacube}{A datacube from the many packages ecosystem.}
}
\value{
A list of lineages that combines agreement area
and agreement action.
}
\description{
Code lineage from agreement titles
}
\examples{
\dontrun{
code_lineage(title = sample(manyenviron::agreements$IEADB$Title, 30))
code_lineage(datacube = manyenviron::agreements)
}
}
