% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{code_memberships}
\alias{code_memberships}
\title{Get conditions and processes to accede memberships}
\usage{
code_memberships(t, condition = NULL, process = NULL)
}
\arguments{
\item{condition}{}

\item{process}{}
}
\value{

}
\description{
Get conditions and processes to accede memberships
}
