% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_agreements.R
\name{code_parties}
\alias{code_parties}
\title{Code Agreement Parties}
\usage{
code_parties(title, memberships = NULL, activity = TRUE)
}
\arguments{
\item{title}{A character vector of treaty titles}

\item{activity}{Do you want the activity of treaty to be coded?
By default, TRUE.}
}
\value{
A character vector of parties
that are mentioned in the treaty title
}
\description{
Identify the countries that are part of an agreement.
}
\details{
The function codes states in treaties alongside,
returning only parties for bileteral treaties
(i.e. 2 parties coded).
The function also returns the "activity" for bilateral
treaties coded, if activity is TRUE.
Bilateral agreements usully detail their
activity and specify area in the last words of the titles.
These last words are abbreviated by the function to
differentiate between bilateral treaties and avoid
false positives being generated since
multiple, different, bileteral treaties are
often signed in the same day.
For the complete list of parties coded please run the
function without and argument (i.e. \code{code_parties()}).
}
\examples{
\donttest{
IEADB <- dplyr::slice_sample(qEnviron::agreements$IEADB, n = 10)
code_parties(IEADB$Title)
}
}
