% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_term.R
\name{code_term}
\alias{code_term}
\title{Code treaty termination date}
\usage{
code_term(title, text = NULL)
}
\arguments{
\item{title}{title column variable.}

\item{text}{treaty text column variable}
}
\value{
a character vector with the term date of the
treaty
}
\description{
Creates a column with the date of termination
of the treaty.
}
\examples{
\donttest{
GNEVAR <- dplyr::slice_sample(manyenviron::agreements$GNEVAR, n = 200)
GNEVAR$Term_dates <- code_term(GNEVAR$Title)
}
}
