% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_term.R
\name{code_term}
\alias{code_term}
\title{Code treaty termination type}
\usage{
code_term(title, text = NULL)
}
\arguments{
\item{title}{title column variable.}
}
\value{
a character vector with the term date of the
treaty and type of termination
}
\description{
Creates a column with the date and type of termination
of the treaty.
}
\examples{
GNEVAR$Term <- code_term(GNEVAR$Title)
}
