% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_agreements.R
\name{condense_agreements}
\alias{condense_agreements}
\title{Condense similar treatyIDs}
\usage{
condense_agreements(database = NULL, idvar = NULL)
}
\arguments{
\item{database}{A "many" package database}

\item{idvar}{Two or more treatyID variables}
}
\value{
A dataframe of treatyID and treatyID references
}
\description{
Different treatyIDs generated for different datasets
might have minor differences in terms of acronym or linkage.
Some minor differences in treatyIDs could mean different treatyIDs
in different datasets actually refer to the same agreement.
The function finds these occurrences and returns the
first treatyID argument entered as a replacement.
}
\examples{
data1 <- data.frame(treatyID = c("CPV-PRT[FSD]_1980A",
"CPV-PRT[FSD]_1990P:FSD_1980A",
"TD06LJ_1981A", "RAMSA_1971A", "WIIEWH_1982P"))
data2 <- data.frame(treatyID = c("TD06LJ_1981A", "RAMSA_1971A",
"WIIEWH_1982P:RAMSA_1971A",
"PRTRPC_1976A", "PRTRPC_1983E1:PRTRPC_1976A"))
condense_agreements(idvar = c(data1$treatyID, data2$treatyID))
}
