% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{fuzzy_agreements_bilateral}
\alias{fuzzy_agreements_bilateral}
\title{Helper function for fuzzy matching bilateral agreements}
\usage{
fuzzy_agreements_bilateral(qID)
}
\arguments{
\item{qID}{qID variable created with \code{qCreate::code_agreements()}}
}
\value{
A data frame with acronyms and qID matches without linkages
}
\description{
Fuzzy match agreements' acronyms for bilateral treaties
}
\details{
Extracts agreement acronyms from qIDs and fuzzy
match their similarities.
See \code{qCreate::code_acronym()} for more details on acronyms.
}
