% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{fuzzy_agreements_bilateral}
\alias{fuzzy_agreements_bilateral}
\title{Helper function for fuzzy matching bilateral agreements}
\usage{
fuzzy_agreements_bilateral(treaty_ID)
}
\arguments{
\item{treaty_ID}{treaty_ID variable created with \code{manypkgs::code_agreements()}}
}
\value{
A data frame with acronyms and treaty_ID matches without linkages
}
\description{
Fuzzy match agreements' acronyms for bilateral treaties
}
\details{
Extracts agreement acronyms from treaty_IDs and fuzzy
match their similarities.
See \code{manypkgs::code_acronym()} for more details on acronyms.
}
