% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articles.R
\name{get_articles}
\alias{get_articles}
\title{Get treaty articles by number or match}
\usage{
get_articles(textvar, article = NULL, match = NULL)
}
\arguments{
\item{textvar}{A text variable}

\item{article}{Either the "preamble" or an article number
across all treaties to be returned.
It can be left NULL if user intends to look for a word in
all treaties.}

\item{match}{A regex match for a word(s) or expression.
For multiple words, please use "|" to divide them.}
}
\value{
A list of treaty sections of the same length
}
\description{
The function splits treaty texts into lists that reflect a structure
based on having a preamble and several articles.
Once articles are split,
you can access all the "preambles",
"membership" or "termination" clauses,
or articles that contain certain word.
}
\details{
If no article or match are declared, only text,
a structured list for each agreement based on articles is returned.
}
\examples{
\donttest{
t <- head(manyenviron::texts$AGR_TXT$Text)
t <- get_articles(t)
get_articles(t, article = "preamble")
get_articles(t, article = "memberships")
get_articles(t, article = "termination")
get_articles(t, article = 1)
get_articles(t, match = "constitution")
get_articles(t, article = "preamble", match = "unofficial")
}
}
