% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articles.R
\name{get_articles}
\alias{get_articles}
\title{Get treaty articles by number or match}
\usage{
get_articles(textvar, article = NULL, match = NULL, treaty_type = "all")
}
\arguments{
\item{textvar}{A text variable}

\item{article}{Would you like to get a specific article?
Null by default.
Other options include the "preamble",
"termination" clause, "membership" clause, "annex",
or an article number.
The specified portion or number
across all treaties will be returned.}

\item{match}{A regex match for a word(s) or expression.
For multiple words, please use "|" to divide them.}

\item{treaty_type}{What types of treaty do you want to look at?
By default, "all".
Other treaty types include:
"agreements", "protocols", "amendments",
"notes", "memorandum", and "resolutions".}
}
\value{
A list of treaty sections of the same length
}
\description{
The function splits treaty texts into lists that reflect a structure
based on having a preamble and several articles.
Once articles are split,
you can access all the "preambles",
"membership" or "termination" clauses,
or articles that contain certain word.
}
\details{
If no article or match are declared, only text,
a structured list for each agreement based on articles is returned.
}
\examples{
\donttest{
t <- sample(manyenviron::texts$AGR_TXT$Text, 30)
get_articles(t)
get_articles(t, article = "preamble")
get_articles(t, article = "memberships")
get_articles(t, article = "termination")
get_articles(t, article = "annex")
get_articles(t, article = 1)
get_articles(t, match = "constitution")
get_articles(t, article = "preamble", match = "amend")
get_articles(t, treaty_type = "agreements")
get_articles(t, treaty_type = "protocols")
get_articles(t, treaty_type = "amendments")
}
}
