% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{get_links}
\alias{get_links}
\title{Get links from treaty titles}
\usage{
get_links(database, dataset, treaty_type = "all")
}
\arguments{
\item{database}{A qPackage database}

\item{dataset}{A qPackage dataset}

\item{treaty_type}{The type of treaties to be returned.
By default, all treaties are returned.
Other options include bilateral or multilateral treaties.}
}
\value{
A dataframe of agreements' qID and their linkages.
}
\description{
Treaties that modify, amend, or expand other treaties
usually specify so in the title.
We use the qIDs generated by \code{code_agreements()} and
\code{condense_qID()} to return a data frame of treaty
links for a database.
}
\examples{
get_links(database = qEnviron::agreements)
get_links(database = qEnviron::agreements, treaty_type = "multilateral")
get_links(dataset = qEnviron::agreements$IEADB, treaty_type = "bilateral")
\dontrun{
samples <- lapply(qEnviron::agreements,
function(x) x[x$Beg > "1991-12-31" & x$Beg < "1993-01-01", ])
migraph::gglineage(get_links(samples))
}
}
