% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_links}
\alias{get_links}
\title{Get links from manyID}
\usage{
get_links(database, dataset, treaty_type = "all")
}
\arguments{
\item{database}{A Package database}

\item{dataset}{A Package dataset}

\item{treaty_type}{The type of treaties to be returned.
By default, all treaties are returned.
Other options include bilateral or multilateral treaties.}
}
\value{
A dataframe of agreements' treatyID and their linkages.
}
\description{
Treaties that modify, amend, or expand other treaties
usually specify so in the title.
We use the manyIDs generated by \code{code_agreements()} and
\code{condense_agreements()} to return a data frame of treaty
links for a database.
}
\examples{
\dontrun{
get_links(database = manyenviron::agreements)
get_links(database = manyenviron::agreements, treaty_type = "multilateral")
get_links(dataset = manyenviron::agreements$IEADB, treaty_type = "bilateral")
samples <- lapply(manyenviron::agreements,
function(x) x[x$Beg > "1991-12-31" & x$Beg < "1993-01-01", ])
migraph::gglineage(get_links(samples))
}
}
