% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_memberships.R
\name{get_memberships}
\alias{get_memberships}
\title{Get memberships' list}
\usage{
get_memberships(country, dataset_id)
}
\arguments{
\item{country}{A country variable}

\item{dataset_id}{A treaty ID variable}
}
\value{
A list of countries part of the treaty
}
\description{
Get memberships' list
}
\examples{
IEADB_MEM$Memberships <- get_memberships(IEADB_MEM$CountryID, IEADB_MEM$IEADB_ID)
}
