% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{get_memberships}
\alias{get_memberships}
\title{Get memberships' list}
\usage{
get_memberships(database, actor, id)
}
\arguments{
\item{database}{A database}

\item{actor}{A actor variable (e.g. country)}

\item{id}{A treaty ID variable}
}
\value{
A dataframe of treaty IDs and actors
part of the treaty
}
\description{
Memberships database have actor column(s) and
treaty column(s) but information on the other
countries that are part to the treaties is often
hard to read.
The function allows to generate a dataframe of
treaty IDs and actors part to the treaty.
}
\examples{
\dontrun{
sample <- manyenviron::memberships$IEADB_MEM
get_memberships(actor = sample$CountryID, id = sample$manyID)
get_memberships(qEnviron::memberships)
}
}
