% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_treaty.R
\name{split_treaty}
\alias{split_treaty}
\title{Split Treaty Texts}
\usage{
split_treaty(textvar)
}
\arguments{
\item{textvar}{text variable}
}
\value{
A structured list for each agreement
}
\description{
Split treaty texts into lists that reflect a structure
based on having a preamble and several articles.
}
\examples{
\donttest{
t <- head(manyenviron::texts$IEADB_TXT$Text)
split_treaty(t)
}
}
