% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{treaty_pdf}
\alias{treaty_pdf}
\title{Convert PDFs to text}
\source{
https://stackoverflow.com/questions/71064939/how-to-check-if-pdf-is-scanned-image-or-contains-text-in-r
}
\usage{
treaty_pdf(path)
}
\arguments{
\item{path}{The file(s) path.
The path can also be specified to a folder with multiple PDF files,
in this case all PDF files in the folder will be converted.}
}
\value{
A list with the converted texts.
}
\description{
Convert PDFs to text
}
\examples{
\dontrun{
treaty_pdf(path = getwd())
}
}
