% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_update.R
\name{update_package}
\alias{update_package}
\title{Updates qPackages}
\usage{
update_package(package = NULL, path = getwd())
}
\arguments{
\item{package}{The name of the package to be updated, optional.
When not declared package name is extracted from description.}

\item{path}{The file path, optional.
If not specified, function get's the working directory.}
}
\value{
Updated files in correct structure for qPackages.
}
\description{
Updates files for qPackages setup with setup_package()
}
\examples{
\dontrun{
update_package("qStates")
}
}
